requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
beautifulsoup4
streamlit
pandas
python-dotenv